Struct nombytes::NomBytes

pub struct NomBytes(_, _);

A wrapper around [bytes::Bytes] to be able to use it with [nom].

Implementations

impl NomBytes

pub fn new(bytes: Bytes) -> Self

Creates a new NomBytes wrapping the provided [Bytes].

Examples

use bytes::Bytes;
use nombytes::NomBytes;

let b = Bytes::new();
let nb = NomBytes::new(b);

pub fn to_str(&self) -> &str

Returns a string slice to the contents of the inner [Bytes].

Examples

use bytes::Bytes;
use nombytes::NomBytes;

let nb = NomBytes::new(Bytes::from("hello"));
assert_eq!(nb.to_str(), "hello");

Panics

Panics if the [Bytes] slice is not UTF-8.

pub fn try_to_str(&self) -> Result<&str, Utf8Error>

Returns a string slice to the contents of the inner [Bytes].

Examples

use bytes::Bytes;
use nombytes::NomBytes;

let nb = NomBytes::new(Bytes::from("hello"));
assert_eq!(nb.try_to_str().unwrap(), "hello");

Errors

Returns Err if the [Bytes] slice is not UTF-8 with a description as to why the provided slice is not UTF-8.

pub fn to_bytes(&self) -> Bytes

Returns a [Bytes] representing the real state of the inner one.

Be careful using a [Bytes] returned from this function to create a new NomBytes for use in the same parsing session, as due to an optimization in [Bytes], creating an empty slice (e.g. asking for the slice of 0..0 or ..0, which [nom] sometimes does) results in a [Bytes] that is unrelated to its source, which causes later offset calculations to give incorrect results.

This behavior is accounted for internally, so as long as you stick to only using NomBytes directly without going to [Bytes] and back, you won’t be affected by this optimization behavior.

Examples

use bytes::Bytes;
use nombytes::NomBytes;

let nb = NomBytes::new(Bytes::from("hello"));
let b = nb.to_bytes();
assert_eq!(b.as_ref(), b"hello");

pub fn into_bytes(self) -> Bytes

Returns a [Bytes] representing the real state of the inner one.

Be careful using a [Bytes] returned from this function to create a new NomBytes for use in the same parsing session, as due to an optimization in [Bytes], creating an empty slice (e.g. asking for the slice of 0..0 or ..0, which [nom] sometimes does) results in a [Bytes] that is unrelated to its source, which causes later offset calculations to give incorrect results.

This behavior is accounted for internally, so as long as you stick to only using NomBytes directly without going to [Bytes] and back, you won’t be affected by this optimization behavior.

Examples

use bytes::Bytes;
use nombytes::NomBytes;

let nb = NomBytes::new(Bytes::from("hello"));
let b = nb.into_bytes();
assert_eq!(b.as_ref(), b"hello");

pub fn into_raw(self) -> (Bytes, Option<RangeType<usize>>)

Returns the values from the inner representation of this type.

See into_bytes for an explanation of why this inner representation exists.

pub fn from_raw((bytes, range_type): (Bytes, Option<RangeType<usize>>)) -> Self

Returns a new NomBytes using the raw values passed in. If these values represent something invalid, you’ll likely see incorrect behavior or even panics. Regular usage should create values using new instead.

See into_bytes for an explanation of why this inner representation exists.